Split up swadgedoku into multiple files
