+++ main/CMakeLists.txt
@@ -44,6 +44,9 @@ idf_component_register(SRCS "asset_loaders/common/heatshrink_encoder.c"
                             "modes/games/picross/picross_select.c"
                             "modes/games/picross/picross_tutorial.c"
                             "modes/games/RoboRunner/roboRunner.c"
+                            "modes/games/swadgedoku/sudoku_data.c"
+                            "modes/games/swadgedoku/sudoku_game.c"
+                            "modes/games/swadgedoku/sudoku_ui.c"
                             "modes/games/swadgedoku/swadgedoku.c"
                             "modes/games/swadgeIt/swadgeIt.c"
                             "modes/music/colorchord/colorchord.c"
